系统计算服务

+++ src/main/java/com/huotu/huobanmall/api/ReportSystem.java
@@ -94,7 +94,7 @@ public interface ReportSystem {
      * @throws Exception
      */
     @RequestMapping(method = RequestMethod.GET)
-    ApiResult salesReport(Output<Long> totalAmount, Output<Float> todayAmount, Output<Float> weekAmount, Output<Float> monthAmount
+    ApiResult salesReport(Output<Float> totalAmount, Output<Float> todayAmount, Output<Float> weekAmount, Output<Float> monthAmount
             , Output<Integer[]> todayTimes, Output<Float[]> todayAmounts
             , Output<Date[]> weekTimes, Output<Float[]> weekAmounts
             , Output<Date[]> monthTimes, Output<Float[]> monthAmounts) throws Exception;
